IControlBoard: Contollers now have double values (0..1) instead of int

+++ projects/RagingBool.Carcosa.Core_cs/Stage/Controller/ControllerModeBase.cs
@@ -97,6 +97,6 @@ namespace RagingBool.Carcosa.Core.Stage.Controller
         }
 
         public abstract void ProcessButtonEventHandler(KeyEventArgs<int, TimedKeyVelocity> e);
-        public abstract void ProcessControllerChangeEvent(ControllerChangeEventArgs<int, int> e);
+        public abstract void ProcessControllerChangeEvent(ControllerChangeEventArgs<int, double> e);
     }
 }
