+++ projects/RagingBool.Carcosa.Core_cs/Stage/Controller/SceneSelectControllerMode.cs
@@ -74,7 +74,7 @@ namespace RagingBool.Carcosa.Core.Stage.Controller
             }
         }
 
-        public override void ProcessControllerChangeEvent(ControllerChangeEventArgs<int, int> e)
+        public override void ProcessControllerChangeEvent(ControllerChangeEventArgs<int, double> e)
         {
             // Nothing to do here...
         }
